services/classifier/requirements.txt
fastapi>=0.116.0
uvicorn[standard]>=0.34.0
pydantic>=2.11.0
pydantic-settings>=2.10.0
python-dotenv>=1.0.1
python-multipart>=0.0.18
requests>=2.32.3
sqlalchemy>=2.0.31
transformers>=4.56.0
tokenizers>=0.22.0
sentence-transformers>=3.4.1
Pillow>=11.0.0
numpy>=2.1.0
opencv-python>=4.11.0.86
ffmpeg-python==0.2.0
qdrant-client==1.9.2
scikit-learn>=1.6.0
httpx>=0.28.1
pytest>=8.2.2
